- Diamondo25's patch for saving 3 whole bytes
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@2474 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/ChatHandler.cpp
@@ -1154,8 +1154,8 @@ void ChatHandler::handleChat(Player *player, PacketReader &packet) {
 void ChatHandler::handleGroupChat(Player *player, PacketReader &packet) {
 	vector<int32_t> receivers;
 	int8_t type = packet.get<int8_t>();
-	uint8_t amount = packet.get<int8_t>();
-	for (size_t i = 0; i < amount; i++) {
+	uint8_t amount = packet.get<uint8_t>();
+	for (uint8_t i = 0; i < amount; i++) {
 		receivers.push_back(packet.get<int32_t>());
 	}
 	string chat = packet.getString();
